src/Services/AI_DraftOrder/requirements.txt
fastapi
uvicorn
python-multipart
requests
chromadb
openai-whisper
# Hoặc thư viện gọi API OpenAI nếu không dùng model local
openai